backend/requirements.txt
# Core LLM and orchestration
openai
langchain
langchain-openai
langchain-core
langchainhub
langgraph

# Vector Store & RAG
faiss-cpu
tiktoken
unstructured
pymilvus
sentence-transformers
torch

# Azure
azure-ai-documentintelligence
azure-core

# ML Agent
scikit-learn
pandas
numpy
openpyxl

# LLM 기반 DST
transformers

# Env & HTTP
python-dotenv
requests
fastapi
uvicorn
pydantic
starlette
itsdangerous

# Dev tools
rich
ipykernel
python-multipart
python-dateutil
python-dotenv

# Database
psycopg2-binary
sqlalchemy
redis
alembic
pgvector
